Add initial version uid crate
This change introduces the initial version of the uid crate. The crate
aids with the creation of unique in-memory IDs.

IDs created by this crate have a couple properties that, depending on
the context, are useful to have:
1) IDs are lightweight and ultimately just a number with some compile
   time magic (and a small creation cost) attached to them.
2) Created IDs are guaranteed to be unique. That is, when created at
   runtime (deserialization is not supported for obvious reasons), two
   newly crated IDs are guaranteed to never be the same.
3) They are fully thread safe. That is, uniqueness is guaranteed even
   among different threads.
4) IDs are immutable. Once created, an ID can be copied, hashed,
   compared, and more, but it cannot be changed.
5) IDs can form name (or rather type) spaces. That is, an ID can be
   parametrized by an arbitrary (potentially private) type and only IDs
   that are parametrized by the same type can interact (i.e., be
   compared etc.).

Being unique, scenarios such as providing a key in a hash table or
database are obvious use cases for IDs produced by this crate.