Use NonZeroUsize as the underlying value
With a recent release of Rust the NonZeroXXX types, which allow for
optimized in-memory representation in conjunction with certain types
(Option being one), have been stabilized. Cases are imaginable where an
Id object is wrapped in an Option and so it can make sense for us to use
such a guaranteed-to-be non-zero value internally. In general having one
bit less available is not a big problem, so the potential benefits
outweigh the downsides.
This patch adjusts the implementation to use NonZeroUsize instead of the
built-in usize as the type holding the internally used ID.