Make crate no_std compatible
By virtue of its non-existent dependency list and using only core
functionality of the standard library, this crate is a perfect candidate
for being a no_std one.
This change adjusts the code to be no_std compatible. Note that tests
still use the std crate, but that should not be a problem.