Implement Default trait for Id struct
This change implements the Default trait for the Id struct. In the past
we have wrestled with ourselves as to whether implementing this trait is
the right thing to do. The concern was that this trait can be seen to
convey the notion of a single "default", i.e., constant initialization.
Yet, the entire point of this crate is that every created ID is
different.
However, ultimately the thinking is that the Default trait mainly
provides the ability to default initialize an object. And in that sense
an Id has a default, as it can be created without additional input from
the environment.