Use latest stable Rust for clippy CI job
The clippy job that we run as part of our continuous integration
pipeline arguably should always use the latest image of Rust stable as
that will have the most up-to-date lints available.
With this change we adjust the CI configuration accordingly. While at
it, annotate the test function in the 'thread_safety' test to prevent
clippy from complaining about our use of collect there -- we want to
collect and only join later to allow for true parallel spawning of
threads.