Declare Id::new_unchecked constructor as unsafe
The Id::new_unchecked constructor relies on the caller upholding certain
invariants or undefined behavior could be the result. As such, it should
have been declared as unsafe.
With this change we adjust it accordingly. While at it, we also make
sure to prevent any undefined behavior from happening in the unlikely
case of overflow of the NEXT_ID counter by asserting that it is
non-zero.