Mark various methods as inline
The Id type comes with a few supporting methods, all of which are small
bodies. With this change we annotate these methods with the #[inline]
attribute to allow for inlining across crate boundaries to happen. Note
that inlining was likely already happening because of our usage of
generics, but with by using this attribute we ensure inlineability
moving forward.