Provide ID types for more integer sizes
The existing Id type unconditionally uses an usize internally, meaning
that it has a certain size. In certain conditions usize may be undesired
and a fixed size integer or one of smaller size may be desired instead.
With this change we introduce ID type variants for the common integer
sizes of 1, 2, 4, and 8 bytes.