Only derive necessary traits in doc tests
The doc tests for the Id type are very liberal when it comes to deriving
traits for the types with which to parametrize Id. That's not
necessarily a good thing.
With this change we only keep the required derives, which has the nice
side effect of better illustrating to users that different traits can be
implemented as necessary for the use-case at hand.