Adjust Debug representation to use tuple formatting
The Debug representation that we use is a bit too verbose, even for
debugging purposes. Specifically, nobody really cares that the type has
an 'id' member.
With this change we emit a tuple representation for debugging purposes,
which just names the type followed by a value, without any additional
clutter.