Bump version to 0.1.6
This change bumps the version of the crate to 0.1.6. The following
notable changes have been made since 0.1.5:
- Adjusted Id::new to panic when a usize counter overflow is
  detected
- Removed T: Copy requirement for Id type
- Annotated Id type with #[repr(transparent)]
- Adjusted Debug representation to use tuple formatting
- Bumped minimum supported Rust version to 1.34

Please note that despite the minimum supported Rust version bump we do
not consider this change backwards compatibility breaking, because
even the new version is in all likelihood still much older than any Rust
version that compiles this crate.